04-auth/samples/stock-agent/server/requirements.txt
agentsts-adk[extensions]>=0.7.17
google-adk[extensions]>=1.25.1
fastapi>=0.110.0
uvicorn>=0.29.0
